requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
coverage==7.6.1
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
python-dotenv==1.0.1
websockets==13.1
yarl==1.13.1
